+++ packages/app-canton/src/operations/signTxn/types.ts
@@ -1,0 +1,28 @@
+import { CantonNetworkType } from '../../proto/generated/canton/common';
+
+export enum SignTxnEvent {
+  INIT = 0,
+  CONFIRM = 1,
+  VERIFY = 2,
+  PASSPHRASE = 3,
+  PIN_CARD = 4,
+}
+
+export type SignTxnEventHandler = (event: SignTxnEvent) => void;
+
+export interface IUnsignedTransaction {
+  protoSerializedPreparedTransaction: string;
+}
+
+export interface ISignTxnParams {
+  onEvent?: SignTxnEventHandler;
+
+  walletId: Uint8Array;
+  derivationPath: number[];
+  txn: IUnsignedTransaction;
+  network?: CantonNetworkType;
+}
+
+export interface ISignTxnResult {
+  signature: string;
+}