+++ packages/app-btc/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cypherock/sdk-app-btc
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [[`e5fb785`](https://github.com/Cypherock/sdk/commit/e5fb78564c335399c09888610f050496b4495480)]:
+  - @cypherock/sdk-utils@0.0.21
+  - @cypherock/sdk-core@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes